requirements.txt
coverage
pre-commit
django
python-dotenv
psycopg2-binary
django-oauth-toolkit
six
requests
pyjwt
djangorestframework
gunicorn
whitenoise
djangorestframework-camel-case
fhir.resources
pydantic
pydantic-core
pyhumps
django-allauth
faker
drf-spectacular
grafana-django-saml2-auth
jsonschema
ipython